feat(context): remove namespace destroy from middleware and enable setting values on context

+++ src/context-middleware-factory/context-middleware-factory.js
@@ -1,24 +1,40 @@
 'use strict';
 
 const continuationLocalStorage = require('cls-hooked');
+const uuid = require('uuid');
 
 class ContextMiddlewareFactory {
   static getMiddleware() {
+    const namespace = this._createNamespace();
+
     return async function(ctx, next) {
-      const namespace = continuationLocalStorage.createNamespace('session');
+      await new Promise(namespace.bind(function(resolve, reject) {
+        namespace.set(
+          'request_id',
+          ctx.request.header['x-request-id'] || uuid.v4()
+        );
 
-      try{
-        await new Promise(namespace.bind(function(resolve, reject) {
-          namespace.set('request_id', ctx.request.header['x-request-id']);
-
-          next().then(resolve).catch(reject);
-        }));
-      } catch(error) {
-        throw error;
-      } finally {
-        continuationLocalStorage.destroyNamespace('session');
-      }
+        next().then(resolve).catch(reject);
+      }));
     };
+  }
+
+  static setOnContext(key, value) {
+    const namespace = this._createNamespace();
+    namespace.set(key, value);
+  }
+
+  static destroyNamespace() {
+    if (this._namespace) {
+      continuationLocalStorage.destroyNamespace('session');
+    }
+  }
+
+  static _createNamespace() {
+    if (!this._namespace) {
+      this._namespace = continuationLocalStorage.createNamespace('session');
+    }
+    return this._namespace;
   }
 }
